feat: implement asynchronous message processing for TTS in voice channel

+++ src/main.py
@@ -15,7 +15,15 @@ load_dotenv()
 ENVIRONMENT = os.getenv("ENVIRONMENT", "development")
 
 intents = discord.Intents.all()
-bot = commands.Bot(command_prefix=".", intents=intents)
+message_queue = asyncio.Queue()
+
+
+class MyBot(commands.Bot):
+    async def setup_hook(self):
+        self.bg_task = asyncio.create_task(process_messages())
+
+
+bot = MyBot(command_prefix=".", intents=intents)
 bot_commands = BotCommands(bot)
 
 
@@ -41,53 +49,64 @@ async def on_command_error(ctx, exception):
     logging.error(exception)
 
 
+async def process_messages():
+    while True:
+        message = await message_queue.get()
+        voice_channel = message.author.voice.channel
+        bot_voice_channel = message.guild.voice_client
+
+        try:
+            if (
+                bot_voice_channel
+                and bot_commands.enabled_to_speak_messages
+                and message.content
+            ):
+                if voice_channel == bot_voice_channel.channel:
+                    tts_audio_path = "tts_audio.mp3"
+                    text_to_say = message.content
+
+                    os.system(
+                        f"gtts-cli '{text_to_say}' --lang {bot_commands.current_language} --output {tts_audio_path}"
+                    )
+
+                    if os.path.exists(tts_audio_path):
+                        bot_voice_channel.play(
+                            discord.FFmpegPCMAudio(tts_audio_path),
+                            after=lambda e: logging.debug("Audio finished"),
+                        )
+
+                        while bot_voice_channel.is_playing():
+                            await asyncio.sleep(1)
+
+                        os.remove(tts_audio_path)
+        except Exception as err:
+            logging.error(err)
+            if ENVIRONMENT == "development":
+                await message.channel.send(err)
+                return
+            await message.channel.send(
+                "An error occurred while trying to speak the message."
+            )
+        finally:
+            message_queue.task_done()
+
+
 @bot.event
 async def on_message(message):
-    if message.author == bot.user :
+    if message.author == bot.user:
         return
 
-    await bot.process_commands(message)
-    
     if message.content.startswith("."):
+        await bot.process_commands(message)
         return
 
-    voice_channel = message.author.voice.channel
-    bot_voice_channel = message.guild.voice_client
-
-    if bot_voice_channel and bot_commands.enabled_to_speak_messages:
-        if voice_channel == bot_voice_channel.channel:
-            try:
-                print(f"{message.author}")
-                tts_audio_path = "tts_audio.mp3"
-                text_to_say =  message.content
-
-                os.system(
-                    f"gtts-cli '{text_to_say}' --lang pt --output {tts_audio_path}"
-                )
-
-                bot_voice_channel.play(
-                    discord.FFmpegPCMAudio(tts_audio_path),
-                    after=lambda e: logging.debug("Audio finished"),
-                )
-
-                while bot_voice_channel.is_playing():
-                    await asyncio.sleep(1)
-
-                os.remove(tts_audio_path)
-            except Exception as err:
-                logging.error(err)
-                if ENVIRONMENT == "development":
-                    await message.channel.send(err)
-                    return
-                await message.channel.send(
-                    "An error occurred while trying to speak the message."
-                )
+    await message_queue.put(message)
 
 
-def main():
+async def main():
     DISCORD_API_TOKEN = os.getenv("DISCORD_API_TOKEN")
-    bot.run(DISCORD_API_TOKEN)
+    await bot.start(DISCORD_API_TOKEN)
 
 
 if __name__ == "__main__":
-    main()
+    asyncio.run(main())
